2022/day09/day09_test.go
package day09_test

import (
	"testing"

	"github.com/henningstorck/advent-of-code/2022/day09"
	"github.com/henningstorck/advent-of-code/common/inputreader"
	"github.com/stretchr/testify/assert"
)

func TestSolve(t *testing.T) {
	tests := []struct {
		filename  string
		expected1 int
		expected2 int
	}{
		{"example01.txt", 13, 1},
		{"example02.txt", 88, 36},
		{"input.txt", 6745, 2793},
	}

	for _, test := range tests {
		reader := inputreader.NewInputReader("../..")
		result1, result2 := day09.Solve(reader, test.filename)
		assert.Equal(t, test.expected1, result1)
		assert.Equal(t, test.expected2, result2)
	}
}
